+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2023, Johann-Mattis List and Robert Forkel'
 author = 'Johann-Mattis List and Robert Forkel'
 
 # The full version, including alpha/beta/rc tags
-release = '2.0.1.dev0'
+release = '2.1.0'
 
 
 # -- General configuration ---------------------------------------------------