+++ src/pyigt/__init__.py
@@ -1,4 +1,4 @@
 from .igt import Corpus, IGT, LGRConformance, Example  # noqa: F401
 from .lgrmorphemes import GlossedWord, GlossedMorpheme  # noqa: F401
 
-__version__ = '2.0.1.dev0'
+__version__ = '2.1.0'